Refactored header into its own file and re-added references
